+++ tests/test_basemaps.py
@@ -1,24 +1,27 @@
 from maplibre.basemaps import (
+    BasemapStyle,
     Carto,
+    OpenFreeMap,
     background,
     construct_basemap_style,
     construct_carto_basemap_url,
-    OpenFreeMap, construct_openfreemap_basemap_url
+    construct_openfreemap_basemap_url,
 )
+from maplibre.layer import Layer, LayerType
+from maplibre.sky import Sky
+from maplibre.sources import RasterSource, RasterTileSource
 
 
-def test_carto_basemaps():
+def test_carto_basemaps() -> None:
     # Act
-    basemap_url = construct_carto_basemap_url(Carto.DARK_MATTER)
+    # basemap_url = construct_carto_basemap_url(Carto.DARK_MATTER)
+    basemap_url = BasemapStyle.carto_url(Carto.DARK_MATTER)
 
     # Assert
-    assert (
-        basemap_url
-        == "https://basemaps.cartocdn.com/gl/dark-matter-gl-style/style.json"
-    )
+    assert basemap_url == "https://basemaps.cartocdn.com/gl/dark-matter-gl-style/style.json"
 
 
-def test_background_style():
+def test_background_style() -> None:
     # Prepare
     color = "white"
 
@@ -29,6 +32,7 @@ def test_background_style():
     # Assert
     assert style["layers"][0]["paint"]["background-color"] == color
 
+
 def test_openfreemap_style() -> None:
     positron = OpenFreeMap.POSITRON
     liberty = "liberty"
@@ -38,3 +42,28 @@ def test_openfreemap_style() -> None:
 
     assert style_url == "https://tiles.openfreemap.org/styles/positron"
     assert construct_openfreemap_basemap_url(liberty) == "https://tiles.openfreemap.org/styles/liberty"
+
+
+def test_basemap_class() -> None:
+    # Act
+    basemap = BasemapStyle(
+        sources={
+            "osm": RasterSource(
+                tiles=["https://a.tile.openstreetmap.org/{z}/{x}/{y}.png"],
+                tile_size=256,
+                attribution="&copy; OpenStreetMap Contributors",
+                max_zoom=19,
+            )
+        },
+        layers=[Layer(id="osm-tiles", type=LayerType.RASTER, source="osm")],
+        sky=Sky(),
+    )
+
+    # Assert
+    style = basemap.to_dict()
+    print(style)
+
+    assert style["version"] == 8
+    assert type(style["sources"]) == dict
+    assert type(style["layers"]) == list
+    assert style["sources"]["osm"]["maxzoom"] == 19