+++ docs/changelog.md
@@ -1,5 +1,16 @@
 # Changelog for MapLibre for Python
 
+## maplibre v0.3.2
+
+* Add `Sky` (#135), `Light` (#137) and `Terrain` (#134) style specifications
+* Add `TerrainControl`
+* Add `RasterDEMSource`
+* Add `RasterSource` as alias for `RasterTileSource`
+* Add `VectorSource` as alias for `VectorTileSource`
+* Add `BasemapStyle` as successor to `construct_basemap_style`
+* Add `BasemapStyle.symbol_layers` to get symbol layers from style
+* Add `BasemapStyle.carto_url` etc as successors to `construct_carto_basemap_url` etc
+
 ## maplibre v0.3.1
 
 * Switch to [MapLibre GL JS v5.3.0](https://github.com/maplibre/maplibre-gl-js/releases/tag/v5.3.0)